Mostly working version of directory checker, now added to unit test.  Replaced erroneous inclusion of astropy.io.fits instead of pyfits.

+++ unit_test/test.sh
@@ -13,44 +13,36 @@ echo ""
 trap '{ echo "ogip_check interrupted;  exiting."; exit; }' INT
 
 
-if [[ -e out ]]; then
-    rm -r out
+if [[ -e out ]]; then rm -r out
 fi
 mkdir out
 
 let errors=0
+let diffcnt=0
 
 echo "##################################################################"
 echo "# Basic tests, i.e., run each file type with no arguments:"
-for file in $( ls inputs/* );
+for file in $( ls inputs );
 do
+#    echo "got file=${file}."
+    if [ -d "inputs/${file}" ]; then 
+	continue
+    fi
     echo ""
-    file=`echo ${file} | awk -F/ '{print $2}'`
     echo "Checking file ${file}"
-    ../ogip_check inputs/${file} &> out/${file}.ogip_check.log
+    ../ogip_check inputs/${file} &> out/${file}.check.log
     retval=$?
     echo "Return status was $retval"
     if [[ "$retval" != "0" ]];  then
        echo "ERROR:  ogip_check returned an error status!"
        let errors+=1
     fi
-    out="$(diff ref/${file}.ogip_check.log out/${file}.ogip_check.log)"
-    #  Bizarre bash way to split output by newlines into array in $diffs
-    IFS=$'\n' read -rd '' -a diffs <<< "$out"
 
-    #  Don't print all of it if it's long:
-    if [[ ${#diffs[@]} != 0 ]]; then
-	if [[ ${#diffs[@]} -lt 5 ]]; then
-	    echo "ERROR:  Differences appear:"
-	    for ln in ${diffs[@]} ; do echo "$ln"; done
-	else
-	    echo "################"
-	    echo "ERROR:  Differences appear;  showing last 5 lines of diff:"
-	    let start=${#diffs[@]}-5
-	    for (( i=${start};i<${#diffs[@]};++i )); do echo -e "${diffs[i]}"; done
-	    echo "################"
-	fi
-	let errors+=1
+    diffs=`diff ref/${file}.check.log out/${file}.check.log`
+
+    if [[ ${#diffs} != 0 ]]; then 
+	echo "ERROR:  Found ${#diffs[@]} differences"
+	let diffcnt+=1
     else
 	echo "No differences."
     fi
@@ -64,7 +56,7 @@ echo "#  Run by specifying the type:"
 echo ""
 file="timing.evt"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} -t TIMING>& out/${file}.ogip_check.log2
+../ogip_check inputs/${file} -t TIMING>& out/${file}.check.log2
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" != "0" ]]; then 
@@ -72,11 +64,11 @@ if [[ "$retval" != "0" ]]; then
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diffs=`diff ref/${file}.ogip_check.log out/${file}.ogip_check.log`
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
 if [[ ${#diffs} != 0 ]]; then
-    echo "ERROR:  Differences appear, for example:"
-    echo diffs | tail
-    let errors+=1
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
 else
     echo "No differences."
 fi
@@ -86,7 +78,7 @@ echo ""
 echo ""
 file="timing_passes.lc"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} -t TIMING >& out/${file}.ogip_check.log2
+../ogip_check inputs/${file} -t TIMING >& out/${file}.check.log2
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" != "0" ]]; then 
@@ -94,14 +86,21 @@ if [[ "$retval" != "0" ]]; then
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/${file}.ogip_check.log2 out/${file}.ogip_check.log2 | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
 
 # SPECTRAL
 echo ""
 file="spectrum.pha"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} -t SPECTRAL >& out/${file}.ogip_check.log2
+../ogip_check inputs/${file} -t SPECTRAL >& out/${file}.check.log2
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" != "0" ]]; then 
@@ -109,28 +108,42 @@ if [[ "$retval" != "0" ]]; then
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/${file}.ogip_check.log2 out/${file}.ogip_check.log2 | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
     
 # RMF
 echo ""
 file="specresp_matrix.rmf"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} -t RMF >& out/${file}.ogip_check.log2
+../ogip_check inputs/${file} -t RMF >& out/${file}.check.log2
 retval=$?
 if [[ "$retval" != "0" ]]; then 
     echo "ERROR:  ogip_check returned an error status!"
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/${file}.ogip_check.log2 out/${file}.ogip_check.log2 | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
 
 # ARF
 echo ""
 file="hexte.arf"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} -t ARF >& out/${file}.ogip_check.log2
+../ogip_check inputs/${file} -t ARF >& out/${file}.check.log2
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" != "0" ]]; then 
@@ -138,14 +151,21 @@ if [[ "$retval" != "0" ]]; then
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/${file}.ogip_check.log2 out/${file}.ogip_check.log2 | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
 
 # CALFILE
 echo ""
 file="fermi_lat_bcf_edisp_calfile.fits"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} -t CALDB >& out/${file}.ogip_check.log2
+../ogip_check inputs/${file} -t CALDB >& out/${file}.check.log2
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" != "0" ]]; then 
@@ -153,7 +173,14 @@ if [[ "$retval" != "0" ]]; then
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/${file}.ogip_check.log2 out/${file}.ogip_check.log2 | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
 
 
@@ -165,8 +192,8 @@ echo "#  Some error cases"
 # Test giving invalid type
 echo ""
 file="fermi_lat_bcf_edisp_calfile.fits"
-echo "Checking file ${file} with bad type"g
-../ogip_check inputs/${file} -t BLAH >& out/bad_type.ogip_check.log
+echo "Checking file ${file} with bad type"
+../ogip_check inputs/${file} -t BLAH >& out/bad_type.check.log
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" == "1" ]]; then 
@@ -176,7 +203,14 @@ else
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/bad_type.ogip_check.log out/bad_type.ogip_check.log | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
 
 
@@ -184,7 +218,7 @@ echo ""
 echo ""
 file="fermi_lat_bcf_edisp_calfile.fits"
 echo "Checking file ${file} with incorrect type"
-../ogip_check inputs/${file} -t TIMING >& out/wrong_type.ogip_check.log
+../ogip_check inputs/${file} -t TIMING >& out/wrong_type.check.log
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" == "1" ]]; then 
@@ -194,7 +228,14 @@ else
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/wrong_type.ogip_check.log out/wrong_type.ogip_check.log | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
 echo ""
 
 
@@ -202,7 +243,7 @@ echo ""
 echo ""
 file="missing.fits"
 echo "Checking file ${file}"
-../ogip_check inputs/${file} >& out/${file}.ogip_check.log
+../ogip_check inputs/${file} >& out/${file}.check.log
 retval=$?
 echo "Return status was $retval"
 if [[ "$retval" == "1" ]]; then 
@@ -212,7 +253,54 @@ else
     let errors+=1
 fi
 echo "Comparing output to reference.  You should see nothing."
-diff ref/${file}.ogip_check.log out/${file}.ogip_check.log | tail
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
+echo ""
+
+
+
+
+
+
+
+#  Test the directory tool
+echo ""
+echo "##################################################################"
+echo "Checking entire contents of directory inputs"
+mkdir out/inputs.logs
+../ogip_check_dir inputs out/inputs.logs >& out/inputs_check.log
+retval=$?
+echo "Return status was $retval"
+if [[ "$retval" != "0" ]]; then 
+    echo "ERROR:  ogip_check_dir returned an error status!"
+    let errors+=1
+fi
+echo "Comparing output to reference.  You should see nothing."
+diffs=`diff ref/${file}.check.log out/${file}.check.log`
+if [[ ${#diffs} != 0 ]]; then
+    echo "ERROR:  Differences appear"
+#    echo ${diffs[@]} | tail
+    let diffcnt+=1
+else
+    echo "No differences."
+fi
+for file in asca_sis_bcf_calfile.fits fermi_lat_bcf_edisp_calfile.fits hexte.arf specresp_matrix.rmf spectrum.pha timing.evt timing_fails.lc timing_passes.lc ; do 
+    diffs=`diff out/${file}.check.log out/inputs.logs/${file}.check.log | tail`
+    if [[ ${#diffs} != 0 ]]; then
+	echo "ERROR:  Differences appear"
+	echo ${diffs[@]} | tail
+	let diffcnt+=1
+    else
+	echo "No differences."
+    fi
+done
+
 echo ""
 
 
@@ -224,10 +312,10 @@ echo ""
 #  Summarize test results
 echo ""
 echo "##################################################################"
-if [[ $errors -eq 0 ]]; then 
-    echo "#  Test completed with $errors errors"
+if [[ $errors == 0 && $diffcnt == 0 ]]; then 
+    echo "#  Test completed with $errors errors and $diffcnt differences"
 else
-    echo "#  ERROR:   Test completed with $errors errors"
+    echo "#  ERROR:   Test completed with $errors errors and $diffcnt differences"
 fi
 echo ""
